+++ custom_components/supernotify/notify.py
@@ -679,7 +679,7 @@ class SupernotifyAction(BaseNotificationService):
         context:
             id: 01HVXT93JGWEDW0KE57Z0X6Z1K
             parent_id: null
-            user_id: e9dbae1a5abf44dbbad52ff85501bb17
+            user_id: a9dbae1a5abf33dbbad52ff82201bb17
         """
         event_name = event.data.get(ATTR_ACTION)
         if event_name is None or not event_name.startswith("SUPERNOTIFY_"):